topic-modelling/TopicModellingAzureFunction/requirements.txt
azure-functions
urllib3
openai
pdfplumber
